[G2P-3477, G2P-3476] Cleaned up g2p_registry and removed social_registry dependency
