+++ g2p_registry/__manifest__.py
@@ -61,13 +61,7 @@
         "g2p_registry_theme",
     ],
     "external_dependencies": {"python": ["python-jose"]},
-    "data": [
-        "security/ir.model.access.csv",
-        "views/main_view.xml",
-        "views/region.xml",
-        "views/registrant_view.xml",
-        "views/res_config_settings_view.xml",
-    ],
+    "data": [],
     "application": True,
     "installable": True,
     "auto_install": False,
