--- g2p_registry/models/__init__.py
@@ -1,3 +1,0 @@
-# from . import ir_module
-from . import region
-from . import registrant
