--- g2p_registry/security/ir.model.access.csv
@@ -1,3 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-g2p_region_admin,Region Admin Access,g2p_registry.model_g2p_region,g2p_registry_base.group_g2p_admin,1,1,1,1
-g2p_region_registrar,Region Registrar Access,g2p_registry.model_g2p_region,g2p_registry_base.group_g2p_registrar,1,0,0,0
