--- g2p_registry/views/main_view.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-   Part of OpenG2P Registry. See LICENSE file for full copyright and licensing details.
--->
-<odoo>
-
-    <record model="ir.ui.menu" id="g2p_registry_base.g2p_main_menu_root">
-        <field name="name">Registry</field>
-    </record>
-
-</odoo>
