--- g2p_registry/views/registrant_view.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-   Part of OpenG2P. See LICENSE file for full copyright and licensing details.
--->
-<odoo>
-    <!-- Individual view -->
-    <record id="view_g2p_registry_individual_region" model="ir.ui.view">
-        <field name="name">g2p.registry.individual.region</field>
-        <field name="model">res.partner</field>
-        <field name="inherit_id" ref="g2p_registry_individual.view_individuals_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='district']" position="after">
-                <field name="region" />
-            </xpath>
-            <xpath expr="//field[@name='registration_date']" position="after">
-                <field name="unique_id" readonly="1" />
-            </xpath>
-        </field>
-    </record>
-
-    <!-- Group view -->
-    <record id="view_g2p_registry_group_region" model="ir.ui.view">
-        <field name="name">g2p.registry.group.region</field>
-        <field name="model">res.partner</field>
-        <field name="inherit_id" ref="g2p_registry_group.view_groups_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//page[@name='basic_info']/group/group" position="inside">
-                    <field name="region" string="Region" />
-            </xpath>
-            <xpath expr="//field[@name='registration_date']" position="after">
-                <field name="unique_id" readonly="1" />
-            </xpath>
-        </field>
-    </record>
-
-    <!-- Group Filter Extensions For Search Criteria -->
-    <record id="view_registry_groups_filter_social_registry" model="ir.ui.view">
-        <field name="name">view.registry.groups.filter.social.registry</field>
-        <field name="model">res.partner</field>
-        <field name="inherit_id" ref="g2p_registry_group.view_registry_groups_filter" />
-        <field name="arch" type="xml">
-            <xpath expr="//search" position="inside">
-                <field name="unique_id" string="Unique ID" filter_domain="[('unique_id', 'ilike', self)]" />
-            </xpath>
-        </field>
-    </record>
-
-     <!-- Individual Filter Extensions For Search Criteria-->
-    <record id="view_registry_individuals_filter_social_registry" model="ir.ui.view">
-        <field name="name">view.registry.individuals.filter.social.registry</field>
-        <field name="model">res.partner</field>
-        <field name="inherit_id" ref="g2p_registry_individual.view_registry_individuals_filter" />
-        <field name="arch" type="xml">
-            <xpath expr="//search" position="inside">
-                <field name="unique_id" string="Unique ID" filter_domain="[('unique_id', 'ilike', self)]" />
-            </xpath>
-        </field>
-    </record>
-
-
-</odoo>
