--- g2p_registry/views/res_config_settings_view.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-   Part of OpenG2P. See LICENSE file for full copyright and licensing details.
--->
-<odoo>
-    <record id="openg2p_inherit_res_config_settings_view_form" model="ir.ui.view">
-        <field name="name">openg2p.res.config.settings.view.inherit.setup</field>
-        <field name="model">res.config.settings</field>
-        <field name="priority" eval="1010" />
-        <field name="inherit_id" ref="base.res_config_settings_view_form" />
-        <field name="arch" type="xml">
-            <app name="g2p_config_settings" position="attributes">
-                <attribute name="string">G2P Registry Settings</attribute>
-            </app>
-        </field>
-    </record>
-</odoo>
