+++ g2p_registry_base/models/registrant.py
@@ -38,6 +38,33 @@ class G2PRegistrant(models.Model):
     occupation = fields.Char()
     income = fields.Float()
     district = fields.Many2one("g2p.district")
+    region = fields.Many2one("g2p.region")
+    unique_id = fields.Char(string="Unique ID", index=True)
+
+    _sql_constraints = [
+        ("unique_id_uniq", "UNIQUE(unique_id)", "unique_id is an unique identifier!"),
+    ]
+
+    @api.model
+    def create(self, vals):
+        record = super().create(vals)
+        if vals.get("is_registrant"):
+            record.generate_unique_id()
+        return record
+
+    def generate_unique_id(self):
+        for rec in self:
+            g2p_que_background_task_model = self.env["g2p.que.background.task"].sudo()
+            if not g2p_que_background_task_model.search(
+                [("worker_payload", "ilike", f'"registrant_id": {rec.id}')]
+            ):
+                g2p_que_background_task_model.create(
+                    {
+                        "worker_type": "id_generation_request_worker",
+                        "worker_payload": {"registrant_id": rec.id},
+                        "task_status": "PENDING",
+                    }
+                )
 
     @api.onchange("phone_number_ids")
     def phone_number_ids_change(self):
