+++ g2p_registry_group/views/groups_view.xml
@@ -19,6 +19,8 @@
                     string="Tags"
                 />
                 <field name="registration_date" />
+                <field name="region" string="Region" />
+                <field name="unique_id" string="Unique ID" />
                 <field name="disabled" column_invisible="1" />
                 <field name="company_id" groups="base.group_multi_company" />
             </tree>
@@ -115,6 +117,7 @@
                             options="{'no_open':True, 'no_create_edit':True,'no_create':True}"
                             readonly="disabled"
                         />
+                        <field name="unique_id" string="Unique ID" readonly="1" />
                     </group>
                     <notebook>
                         <page string="IDs" name="ids">
@@ -197,6 +200,7 @@
                                     <div class="o_address_format">
                                         <field name="address" readonly="disabled" />
                                     </div>
+                                    <field name="region" string="Region" />
                                 </group>
                                 <group colspan="2">
                                     <field
