+++ g2p_registry_individual/views/individuals_view.xml
@@ -17,6 +17,7 @@
                 />
                 <field name="birthdate" />
                 <field name="registration_date" />
+                <field name="unique_id" readonly="1" />
                 <field name="disabled" column_invisible="1" />
                 <field name="company_id" groups="base.group_multi_company" />
             </tree>
@@ -125,6 +126,7 @@
                     </div>
                     <group colspan="4" col="4">
                         <field name="registration_date" readonly="disabled or registration_date" />
+                        <field name="unique_id" readonly="1" />
                         <label for="lang" invisible="active_lang_count &lt;= 1" />
                         <div class="o_row" invisible="active_lang_count &lt;= 1">
                             <field name="lang" readonly="disabled" />
@@ -147,6 +149,8 @@
                                         readonly="disabled"
                                     />
                                     <field name="district" string="District" />
+                                    <field name="region" string="Region" />
+
                                 </group>
                                 <group colspan="3">
                                     <field
